Don't pad unused terms within vector commitments
Arguably, this makes the code variable-time to a secret (the amount of
commitment terms used). Since it's the amount of terms set to any value
(including zero), and the code to pad each commitment was already variable
time, it should be fine. It just cements the amount of variables within a
commitment is part of the program (not part of the witness), and its their
values which are secrets.