Support proving proofs with empty `aL`, `aR` again within `generalized-bulletproofs`
Regression introduced when padding was removed. 'empty' proofs are mapped to
`n = 1` proofs (as the smallest set of generators allowed is for `n = 1`, and
a `n = 0` proof is undefined).

These proofs technically have use for anyone who wants to prove a series of
commitments/constrain them, but never needs to calculate any products within
the witness.